video-service/requirements.txt
# Python deps for video-service (Remotion renders + FastAPI + RQ).
# Remotion CLI: Node + video-templates (see Dockerfile / Makefile).

fastapi==0.115.6
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
httpx==0.27.2
anthropic==0.40.0
supabase==2.10.0
redis==5.2.1
rq==2.1.0
pydantic==2.10.3
pydantic-settings==2.6.1
